Fixed problem with exporters
If Nokia-Exporter was running, the normal VCard-Export did not work anymore.
Therefore the module_vcard.php is now also modified and added to this repo.
I also added the changed template for the contactlist, so the export for
Fritz-Box and for Nokia is also visible in the frontend.
The actions.php had to be modified therefore as well.